requirements/dev.txt
# Development requirements
-r base.txt

# Debug and Development Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0
ipython>=8.15.0
ipdb>=0.13.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.3.0
freezegun>=1.2.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Profiling and Performance
django-silk>=5.0.0
memory-profiler>=0.61.0


whitenoise>=6.5.0